Add extended password digest functionality to addUserToken
An optional parameter, $passwordDigestExt, has been added to the addUserToken function. This allows the password digest to be extended by base64 encoding it with sha1. The passwordDigestExt option is non-standard and should only be used for interop issues where the message receiver desires an extra SHA-1 Hash of the password.